requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
alembic
python-dotenv
email-validator
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings
python-multipart